+++ c-based-ffi/src/rust_api.h
@@ -6,11 +6,11 @@ extern uint32_t r_add(uint32_t, uint32_t);
 
 typedef void (*callback)(uint32_t);
 
-typedef void (*FnCommit)(uint8_t);
+typedef void (*FnCommit)(uint16_t);
 
 typedef void (*FnVoid)();
 
-typedef void (*FnSetState)(char *, int8_t **, uint8_t *, size_t);
+typedef void (*FnSetState)(char *, int16_t **, uint16_t *, size_t);
 
 extern void r_run_callbacks(FnCommit, FnVoid, FnVoid, FnSetState);
